* Initial GrinJoin Support * Unicode path support * Payment proofs part 1 * Partial v3 slate support * TxPool performance improvements * JSON-RPC Wallet APIs * Fixing critical tx aggregation sorting bug

+++ src/Crypto/RandomNumberGenerator.cpp
@@ -21,7 +21,7 @@ SecureVector RandomNumberGenerator::GenerateRandomBytes(const size_t numBytes)
 {
 	SecureVector buffer(numBytes);
 #ifdef _WIN32
-	const NTSTATUS status = BCryptGenRandom(nullptr, buffer.data(), buffer.size(), BCRYPT_USE_SYSTEM_PREFERRED_RNG);
+	const NTSTATUS status = BCryptGenRandom(nullptr, buffer.data(), (ULONG)buffer.size(), BCRYPT_USE_SYSTEM_PREFERRED_RNG);
 	if (!BCRYPT_SUCCESS(status))
 	{
 		throw CryptoException("RNG Failure");
